Setting up for ETL.
This commit sets up the parser, and performance metrics computation for the overall project. Introduced the basic structure for the nyctlc_ingest module which will ingest the required data, perform requie
d transformations and store the data at the target locations. Also introduced the config.py file, that acts as the configurations file for the project.