Added the structured function calls for the the two ETL jobs.
Updated the nyctlc_ingest module to update the core module to add the functions - pickup_zones_by_passengers and pickup_borough_by_rides to take care of the two ranking jobs. Updated the utils file in the nyctlc_ingest to add the support functions for the ETL procedures in the core file. Updated the parser to return parsed arguments.